+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Holidays for Ireland.
+
 ## [0.4.0] - 2020-11-02
 
 ### Added